+++ source_gen/source_gen/lib/source_gen.dart
@@ -8,7 +8,7 @@ export 'src/builder.dart';
 export 'src/constants.dart' show ConstantReader;
 export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';
-export 'src/library.dart' show LibraryReader;
+export 'src/library.dart' show AnnotatedElement, LibraryReader;
 export 'src/revive.dart' show Revivable;
 export 'src/span_for_element.dart' show spanForElement;
 export 'src/type_checker.dart' show TypeChecker;